I2C Reset Path: Add Poll of SCL High Before Issuing Stop Command
At the end of the I2C reset sequence, a 'stop' command is sent to
each port on the master that was reset.  Before sending this 'stop'
command, a poll has been added such to check that SCL (clock line)
is high. This addresses a known issue with current TPMs.

Change-Id: I13b199f5b4d10d7224d3ccc5d1a8baaecb4f2326
CQ: SW405816
Backport: release-fips910
Reviewed-on: http://ralgit01.raleigh.ibm.com/gerrit1/52636
Reviewed-by: Christopher J. Engel <cjengel@us.ibm.com>
Reviewed-by: Timothy R. Block <block@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Reviewed-by: Nicholas E. Bofferding <bofferdn@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Reviewed-by: William G. Hoffa <wghoffa@us.ibm.com>