Error when hb tries to access reserved memory past limit
An issue has recently occurred when hb reserved memory past
the allowed limit (256M-4K) and overwrote some of the PHYP
memory, which caused bugs that were difficult to debug. This
change is to check the memory address we are trying to reserve
and throw an error if the address is at or above the allowed
limit. We only execute this check if PHYP is running and only
on components not belonging to PHYP (PHYP components may be
placed outside of the limit).

Change-Id: Ic62a7b724abc3b29b7872d0af47de8c68cde2ea8
RTC:186332
Reviewed-on: http://ralgit01.raleigh.ibm.com/gerrit1/52850
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Reviewed-by: Nicholas E. Bofferding <bofferdn@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Reviewed-by: Michael Baiocchi <mbaiocch@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Reviewed-by: Daniel M. Crowell <dcrowell@us.ibm.com>