Populate the Severity of PLAT Errors
An issue was observed where recovered PLAT error logs were not
being properly hidden/skipped and incorrect severity was
assigned to those error logs. The reason is that the path of
code that PLAT errors take fails to overwrite the default severity
(unrecoverable). A fix is to explicitly set the severity of all
errors at the end of rcToErrl.

Change-Id: Ib0aff72a800e0ed7a5b48e753049d2f3422406b4
CQ:SW410923
Reviewed-on: http://ralgit01.raleigh.ibm.com/gerrit1/53221
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Reviewed-by: Christian R. Geddes <crgeddes@us.ibm.com>
Reviewed-by: Prachi Gupta <pragupta@us.ibm.com>
Reviewed-by: Daniel M. Crowell <dcrowell@us.ibm.com>