Set PAYLOAD_KIND attribute based on fused bits in scom register
The scom register 0xF000F, bit 57 tells us whether or not the fuse
has been blown on this processor. If it is set, that means that it
is up to hostboot to set the PAYLOAD_KIND attribute. We decide this
based on if we are in SMT4 or SMT8 mode, which is known during istep 6.
If we are in SMT4 and the fuse is blown, we can assume PAYLOAD_KIND is
OPAL, if we are in SMT8 and the fuse is blown, we are booting PHYP. In
a lab scenario on which a system has blown fuses, and we want to force
OPAL boot on a system we are forcing to be SMT8, we would have to apply
an attribute override to make ATTR_PAYLOAD_KIND = 0x2 . 0x2 is the enum
for SAPHHIRE as the payload kind.

Change-Id: Ia15dee4dd67130691440bf7fe76915acfb81a19d
RTC: 186965
Reviewed-on: http://ralgit01.raleigh.ibm.com/gerrit1/52957
Reviewed-by: Martin Gloff <mgloff@us.ibm.com>
Reviewed-by: ILYA SMIRNOV <ismirno@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Reviewed-by: Matt Derksen <mderkse1@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Reviewed-by: Daniel M. Crowell <dcrowell@us.ibm.com>