Remove Istep 10.3 From IPL Flow
Istep 10.3 (host_set_voltages) has been moved to Istep 8.12
and needs to be no-oped. Executing host_set_voltages twice
causes crashes due to secureboot thinking we are trying to
skip an istep. Note that validateSecuritySettings that was
performed in 10.3 has now been moved to 10.4 (with the required
sub-routines).

Change-Id: I81284157dedebb3f4ee357ce28b29b1dd6a3fe8a
CQ:SW416209
CMVC-Coreq:1044932
Backport: release-fips910
Reviewed-on: http://ralgit01.raleigh.ibm.com/gerrit1/53443
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Reviewed-by: Nicholas E. Bofferding <bofferdn@us.ibm.com>
Reviewed-by: Michael Baiocchi <mbaiocch@us.ibm.com>
Reviewed-by: Marshall J. Wilks <mjwilks@us.ibm.com>
Reviewed-by: Daniel M. Crowell <dcrowell@us.ibm.com>