Always set RISK_LEVEL regardless of value
Fixing a bug where we only set ATTR_RISK_LEVEL if the value is
non-zero.  This leads to cases where we cannot clear the value
once we set it.
I also changed the persistency of ATTR_RISK_LEVEL so that it
doesn't persist across boots since it will always be set to
a new value during each boot.

Change-Id: I9df1e505755f860524a3b1af46c68f315f142066
CQ: SW415934
Reviewed-on: http://ralgit01.raleigh.ibm.com/gerrit1/53451
Reviewed-by: Dean Sanner <dsanner@us.ibm.com>
Reviewed-by: Christian R. Geddes <crgeddes@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Reviewed-by: Daniel M. Crowell <dcrowell@us.ibm.com>