Check TYPE_NODE Targets Twice for Serial/Part Number Updates
There was an issue where the system serial number was not
updating properly in hostboot. The problem is that changing
just the system serial/part number does not trigger FRU update on
our side. The fix is to add an explicit check for SN/PN in
the code path that compares SEEPROM and PNOR records/keywords
for TYPE_NODE targets only.

Change-Id: I3c4d43e2ca9ade0e03c8f37198b67dd7e7efbbde
CQ:SW414411
Backport:release-op910
Reviewed-on: http://ralgit01.raleigh.ibm.com/gerrit1/53754
Reviewed-by: Nicholas E. Bofferding <bofferdn@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Reviewed-by: Michael Baiocchi <mbaiocch@us.ibm.com>
Reviewed-by: Marshall J. Wilks <mjwilks@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Reviewed-by: Daniel M. Crowell <dcrowell@us.ibm.com>