Fix rollover of PLID numbers
Need to avoid incrementing into the processor node
field of the PLID.

Decided to use the highest bit in the last 3 bytes as a dividing bit.
[0x##800000 - 0x##FFFFFF] -- for pre-boot setup errors.
[0x##000000 - 0x##7FFFFF] -- for post setup errors

When setting up from PNOR, we will look for the max EID in the post-setup range.
If no max is found, restart at 0.

Change-Id: I4a8811d9c93039090504ca9cf8c0eebc6812a4ca
CQ:SW416592
Reviewed-on: http://ralgit01.raleigh.ibm.com/gerrit1/54362
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Reviewed-by: Daniel M. Crowell <dcrowell@us.ibm.com>