Fixed the SBE FFDC parsing issue of mssing traces and attributes
Not exactly sure why the code was not printing out the wanted traces and
attribute data, but apparently the files that contains the code for the SBEIO
FFDC parser, file sbeIoErrlUd.H and file sbeIoFfdcType.H, are a copy of what
is in the FSP repo, files sbeUserDataParser.C and sbeFFDCType.H.  The HB
files were out of sync with FSP files.  Once I updated the HB files, literally
just copied the contents of the FSP files into the HB files, the traces and
attribute data appeared.  Turns out it was just a syncing issue.

Change-Id: I84e55d0cd0ed81034bbf774af049f5b486278f5f
CQ:SW415480
Reviewed-on: http://ralgit01.raleigh.ibm.com/gerrit1/54057
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Reviewed-by: ILYA SMIRNOV <ismirno@us.ibm.com>
Reviewed-by: Daniel M. Crowell <dcrowell@us.ibm.com>
Reviewed-on: http://ralgit01.raleigh.ibm.com/gerrit1/58001
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>