Remove if that was catching SBE chipop err logs and forcing reboot
Before Hostboot had the ability to parse SBE FFDC, we would perform
a special shutdown that would request the FSP to parse the SBE FFDC
in the event of a SBE chip-op (PSU or FIFO) failure. Now that hostboot
can do this we do not need to go down this path in this case. The
only time we should do this is if the SbeRetryHandler has decided
that this is required.

CQ: SW417255
Change-Id: I8f52d52294d84c9e10faf7d8bc500808463aa57f
Reviewed-on: http://ralgit01.raleigh.ibm.com/gerrit1/54063
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Reviewed-by: Nicholas E. Bofferding <bofferdn@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Reviewed-by: Daniel M. Crowell <dcrowell@us.ibm.com>