Pass i_skipComm to _buildOccs
In the _resetOccs path the i_skipComm parameter is not getting
passed into _buildOccs function that (almost) unconditionally
executes sendOccPoll, regardless of whether the comm was supposed
to be skipped or not. That 1) defeats the purpose of i_skipComm,
and 2) may not work in the OCC reset path depending on how far into
OCC setup the code has gotten before the reset was requested.

Change-Id: I044e1193d8ddf022ada902030c74c75076af2f42
Reviewed-on: http://ralgit01.raleigh.ibm.com/gerrit1/54522
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Reviewed-by: Marshall J. Wilks <mjwilks@us.ibm.com>
Reviewed-by: Christopher J. Cain <cjcain@us.ibm.com>
Reviewed-by: Andres A. Lugo-Reyes <aalugore@us.ibm.com>
Reviewed-by: Nicholas E. Bofferding <bofferdn@us.ibm.com>
Reviewed-by: Daniel M. Crowell <dcrowell@us.ibm.com>