Fixed both NIMBUS and CUMULUS. They are now making the call to mss_thermal_init
Neither NIMBUS or Cumulus was making the call to mss_thermal_init.  For
Cumulus, the issue was trying to create a target list by getting a
list of chiplets when MEMBUF is a chip.  For NIMBUS, well, the code
was not even there.  For Nimbus a target list for the MCS, which is
chiplet in this case, is created and used during the call.

Change-Id: I17c4e1b12435c4dcfaa528b905d67a59962b3bca
CQ: SW416904
Reviewed-on: http://ralgit01.raleigh.ibm.com/gerrit1/54621
Reviewed-by: Christian R. Geddes <crgeddes@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Reviewed-by: Daniel M. Crowell <dcrowell@us.ibm.com>