Update to p9_xip_tool to handle stand-alone DDCO ring files.
In this update, we're adding support to p9_xip_tool to be able
to dissect and extract stand-alone DDCO ring files, i.e. files
that are DD packaged through using the dd_container API and
which is the DD packaging that XIP "understands".

Right now, dissect and extract (of a specific DD level) can
only be achieved if the DDCO binary is attached to the XIP image
as one of the rings sections, .rings, .overlays or .overrides.

Key_Cronus_Test=XIP_REGRESS

Change-Id: I37acc7f97b175577bbda97fd542ff24f60c46fb6
Original-Change-Id: I8be00742f05b4df61652a1de6d0230bb5d203f97
Reviewed-on: http://ralgit01.raleigh.ibm.com/gerrit1/54704
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: HWSV CI <hwsv-ci+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Tested-by: Cronus HW CI <cronushw-ci+hostboot@us.ibm.com>
Tested-by: PPE CI <ppe-ci+hostboot@us.ibm.com>
Reviewed-by: Sumit Kumar <sumit_kumar@in.ibm.com>
Reviewed-by: Richard J. Knight <rjknight@us.ibm.com>
Reviewed-by: Jennifer A. Stofer <stofer@us.ibm.com>