bootconfig: Describe pnor boot flags
The reserved boot flag is now used by a BMC to signal to the host
firmware that it supports the mbox protocol.

Currently this has no affect on Hostboot, but is used by skiboot.
In the future hostboot could chose to use the mbox protocol to read and
write PNOR.

Change-Id: I51bc7878eabc975954371ada838d0a348b12ef9d
Signed-off-by: Joel Stanley <joel.stanley@au1.ibm.com>
Reviewed-on: http://ralgit01.raleigh.ibm.com/gerrit1/54740
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Reviewed-by: Daniel M. Crowell <dcrowell@us.ibm.com>