Re-order sbex calls in presimsetup to get paths updated correctly
We added another sbex call for the track 1045730 which changes the
simicsInfo file. This sbex getting called after we had done some
additions to that file, which basically wiped out our additions.
These additons we made were making simics look at the correct
path when determining which pnor to use. Without this simics was
using the nimbus.pnor image from the backing build, which does
not have tests built in

Change-Id: I73a95ec375aa8d46a63af8e900ccefde06b073a8
Reviewed-on: http://ralgit01.raleigh.ibm.com/gerrit1/54790
Reviewed-by: ILYA SMIRNOV <ismirno@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Reviewed-by: Christian R. Geddes <crgeddes@us.ibm.com>