Update autocitest to collect all hostboot dump info prior to failure
At the end of the autocitest script, which controls how our CI runs,
we were failing if any of the commands to dump various info about
the simics run happened to fail. We would prefer to have continue
trying to gather information about the run and fail at the end
if one of the "success" requirements has not been achieved.

Change-Id: Ic7b9d71ae6be9263f45a8fbe4a79cd35c99ac690
Reviewed-on: http://ralgit01.raleigh.ibm.com/gerrit1/54810
Reviewed-by: ILYA SMIRNOV <ismirno@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Reviewed-by: William G. Hoffa <wghoffa@us.ibm.com>
Reviewed-by: Daniel M. Crowell <dcrowell@us.ibm.com>