PRD: DMD address translation bug.
First issue is that page garding is completely broken because we were
using the wrong fields for the bank and bank group.

Second issue is for MCAs that only have one DIMM plugged and that DIMM
only has one rank of memory. In this case, rank/DIMM/port deallocation
will deconfigure a range that is twice as big as the actual range.

Change-Id: Ieb615260fe7ba6adb6a68ce3d4e3e24076351e72
CQ: SW419378
Reviewed-on: http://ralgit01.raleigh.ibm.com/gerrit1/55221
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Reviewed-by: Benjamin J. Weisenbeck <bweisenb@us.ibm.com>
Reviewed-by: Caleb N. Palmer <cnpalmer@us.ibm.com>
Reviewed-by: Brian J. Stegmiller <bjs@us.ibm.com>
Reviewed-by: Zane C. Shelley <zshelle@us.ibm.com>
Reviewed-on: http://ralgit01.raleigh.ibm.com/gerrit1/55360
CI-Ready: Zane C. Shelley <zshelle@us.ibm.com>