ATTR_PBAX_GROUPID: add global tag
PBAX_GROUPID is a proc level attr that needs
to be incrememnted per node. The only way to
set this correctly in MRW is to add a global
tag. However, making it global may break other
systems as the global value in their respective
MRW won't be set. It will be set on proc_socket
level. Therefore, added a check to only copy
global values if they are not empty (aka someone
must explicitly set the global value for us to
use those values).

Change-Id: I3055d8296c010f996d98d7578a9adee5b763d316
Reviewed-on: http://ralgit01.raleigh.ibm.com/gerrit1/55257
Reviewed-by: Santosh S. Puranik <santosh.puranik@in.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Reviewed-by: William G. Hoffa <wghoffa@us.ibm.com>