Force single-threaded access to HWPs in PRD
Hostboot does not support thread local storage, however several
HWPs use this concept.  Specifically the memory HWPs use it for
their c_str() function that prints out target strings.  If these
HWPs are called in multiple threads there is a chance of a buffer
overrun that leads to a crash.

The fix is to create a new INVOKE-style macro that befores the
same mutex lock as a regular FAPI_INVOKE_HWP call.

Change-Id: I550fc752f6a6b1bb51ecb4f9c8a7011f98bb8cd3
CQ: SW420212
Reviewed-on: http://ralgit01.raleigh.ibm.com/gerrit1/55289
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Reviewed-by: Daniel M. Crowell <dcrowell@us.ibm.com>
Tested-by: Daniel M. Crowell <dcrowell@us.ibm.com>