Do not elevate severity of reconfig error log
The code as currently written will elevate the severity of any
log that is used to terminate the hostboot ipl.  In most cases
this is the correct behavior as we always want a visible log.
However, there are cases where hostboot terminates for the sole
purpose of triggering a reconfig loop via the FSP in order to
recover from a recoverable hardware issue.  In this case the log
shouldn't be visible since we don't want the customer to take
any actions.

The fix is to add a little bit of logic to differentiate these
two scenarios to control the severity setting.

Change-Id: I7253aec8c28a40c5cdebf4933ceccbecd119b9f4
CQ: SW420495
Backport: yes
Reviewed-on: http://ralgit01.raleigh.ibm.com/gerrit1/55379
CI-Ready: Nicholas E. Bofferding <bofferdn@us.ibm.com>
CI-Ready: Daniel M. Crowell <dcrowell@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Reviewed-by: Nicholas E. Bofferding <bofferdn@us.ibm.com>
Reviewed-by: Marshall J. Wilks <mjwilks@us.ibm.com>
Reviewed-by: William G. Hoffa <wghoffa@us.ibm.com>