Secure Boot: Mark redundant TPM not present until SMP is enabled
Until the SMP is up, a redundant TPM is not able to be presence detected.  This
change updates the TPM presence detect routine to report a TPM connected to
remote processor as not present when the remote processor is not connected to
the SMP, so that Hostboot doesn't hit errors attempting to access/use the TPM at
inappropriate times.

Change-Id: I10aa683ec073be2b44e810746bee320dd6366184
RTC: 184515
Reviewed-on: http://ralgit01.raleigh.ibm.com/gerrit1/55814
Reviewed-by: ILYA SMIRNOV <ismirno@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Reviewed-by: Christopher J. Engel <cjengel@us.ibm.com>
Reviewed-by: Michael Baiocchi <mbaiocch@us.ibm.com>
Reviewed-by: Daniel M. Crowell <dcrowell@us.ibm.com>