Use SCOM to access Master Processor's SBE MSG register
Previously we always used CFAM to acess the sbe_msg_register. However,
this will fail if we attempt to read the master via CFAM. The master processor
cannot call getCfam on itself. In the case where we need to get this msg reg
from the master we must use scom.

RTC: 180604
CQ: SW421405
Change-Id: Iaa199361272e30b73114fafb8d3f6887c794464b
Reviewed-on: http://ralgit01.raleigh.ibm.com/gerrit1/55883
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: HWSV CI <hwsv-ci+hostboot@us.ibm.com>
Reviewed-by: Jennifer A. Stofer <stofer@us.ibm.com>
Reviewed-on: http://ralgit01.raleigh.ibm.com/gerrit1/55893
Reviewed-by: Christian R. Geddes <crgeddes@us.ibm.com>