Support multiple nodes in HBRT - Remove Single Node Items
There are some items that applied to a single node implementation that can
be removed after all dependencies on them have been removed by other tasks.
This single node support is being moved/removed.  Also a check is being
added to the VPD support so the node value does not index beyond the end
of the array of reserved memory addresses.

Change-Id: I7fea5ac9cdfa9f37caa57905d320633a8039e79d
RTC: 186585
Reviewed-on: http://ralgit01.raleigh.ibm.com/gerrit1/55917
Reviewed-by: Christian R. Geddes <crgeddes@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Reviewed-by: Prachi Gupta <pragupta@us.ibm.com>
Reviewed-by: William G. Hoffa <wghoffa@us.ibm.com>