Completely hide attributes that have no value
Serverwiz2 continues to output invalid attribute syntax where
attributes are defined with blank default tags like this.
	<attribute>
		<id>VCS_AVSBUS_BUSNUM</id>
		<default></default>
	</attribute>
Which then get converted into something like this in our
parsed xml.
	<attribute>
		<id>VCS_AVSBUS_BUSNUM</id>
	</attribute>

That syntax results in our xmltohb parser using a value of zero
instead of the defined default in the attribute definition since
it looks like the value is being explicitly overridden.

The change here is to completely skip adding any attribute that
does not have a value filled in.  The attribute's value will then
be sourced from the defaults in attribute_types/target_types.

Change-Id: I3565ff3bf50f9cd8b8f5ef108d8690016cf3f5f2
Reviewed-on: http://ralgit01.raleigh.ibm.com/gerrit1/55970
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Reviewed-by: Prachi Gupta <pragupta@us.ibm.com>
Reviewed-by: Christian R. Geddes <crgeddes@us.ibm.com>
Reviewed-by: William G. Hoffa <wghoffa@us.ibm.com>