Modify debug framework to be build-independent
During boot, Hostboot will push key pointers into memory.  This
allows the debug tools to find the pointers (using a known static
memory address) to base memory accesses on.  This replaces the
existing symbol lookup that we use now.  That means we don't
need to have the exact symbol file for the build we're debugging
against.

Change-Id: I4618e15a3dc90acc3a89520a502eb818c1b4258c
Reviewed-on: http://ralgit01.raleigh.ibm.com/gerrit1/58063
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Reviewed-by: Daniel M. Crowell <dcrowell@us.ibm.com>