Update hreset HWP to always use scoms during HBRT
Previsouly there were checks in p9_sbe_hreset that made sure that
we only did getcfam ops on slave chips and did scoms to access the
registers on the master chips. This is correct for hostboot during
IPL time but during runtime we always want to use scom to access
these registers. This is safe because the fabric us up and we can
perform xscoms on both chips.

Change-Id: I5c48053d90278e1b3f9aacd5a74dda4ddc3620bb
RTC: 190058
CQ: SW422692
Reviewed-on: http://ralgit01.raleigh.ibm.com/gerrit1/56183
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Sachin Gupta <sgupta2m@in.ibm.com>
Reviewed-by: SRINIVAS V. POLISETTY <srinivan@in.ibm.com>
Reviewed-by: Jennifer A. Stofer <stofer@us.ibm.com>