Remove logic to deconfigure PHBs in higher position within a pec
Current deconfig logic will deconfigure PHBs with a higher
unit position under a given pec, if we were deconfiguring
the PHB with a lower unit position. For example, PEC1
has PHB1 and PHB2. If we deconfigured PHB1, then we
would by association deconfigure PHB2 as well. However,
if we deconfigured PHB2, then we would not deconfigure PHB1.
This commit removes the logic where if PHB2 is deconfigured,
PHB1 is deconfigured as well. To summarize, the end result
is to be able to deconfigure PHB1 or PHB2 with no effect
on the other. HWPs will handle the hw limitations we have between PHBS.

Change-Id: I95b31f4d051c15d896d18adf0a7f7c1b994f5928
CQ: SW417485
Reviewed-on: http://ralgit01.raleigh.ibm.com/gerrit1/56499
Reviewed-by: Martin Gloff <mgloff@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Reviewed-by: William G. Hoffa <wghoffa@us.ibm.com>
Reviewed-by: Roland Veloz <rveloz@us.ibm.com>
Reviewed-by: Daniel M. Crowell <dcrowell@us.ibm.com>