attrrp_rt.C: translateAddr returns input address by default
After a HBRT reset, PEER_TARGET attribute had its value
preserved from prior to the reset. This attribute is a
target pointer, so its address value needs to be
translated for HBRT. When the translation is attempted
again after the reset, we have a valid address already.
The second translation fails and returns NULL. This fix
was to return the original address instead of NULL
when the translation isn't possible.

CQ:SW422523
Change-Id: I75de8d1e780140ca1979dd6fdf237c6c0526f08b
Reviewed-on: http://ralgit01.raleigh.ibm.com/gerrit1/56332
Reviewed-by: Richard J. Knight <rjknight@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Reviewed-by: Martin Gloff <mgloff@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Reviewed-by: Christian R. Geddes <crgeddes@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Reviewed-by: William G. Hoffa <wghoffa@us.ibm.com>