Ignore dummy files when parsing error log data
Hit a problem where the patch process was leaving around files
ending in .orig that would be noticed by our error log data
generator script.  The script then reports a duplicate id and
fails the build.

The fix is to only look at files that end in H or h.

Change-Id: I816c080eb6690075cedce7c5f13c9c40a1af3d93
Reviewed-on: http://ralgit01.raleigh.ibm.com/gerrit1/56479
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Reviewed-by: Corey V. Swenson <cswenson@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Reviewed-by: Richard J. Knight <rjknight@us.ibm.com>
Reviewed-by: Daniel M. Crowell <dcrowell@us.ibm.com>