Secure Boot: Inform FSP of key transition progress
The existing key transition flow updates all SBE SEEPROMs within the system in
the same boot, thereby quiescing all SBEs.  This compromises the ability of FSP
to detect a key transition TI at the end of the update process because the SBE
is not alive to service FSP's incoming SBE FIFO request to read the memory
holding the attention information.  This change adds new support to log
the key transition progress in a node target attribute introduced in an
earlier commit and send that status to FSP via a new mailbox message.

Change-Id: I828184ae8be89ac87137d7510a3c375cef2e05d7
CQ: SW418697
CMVC-Prereq: 1053806
CMVC-Prereq: 1053552
Reviewed-on: http://ralgit01.raleigh.ibm.com/gerrit1/56677
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Reviewed-by: Michael Baiocchi <mbaiocch@us.ibm.com>
Reviewed-by: Marshall J. Wilks <mjwilks@us.ibm.com>
Reviewed-by: Daniel M. Crowell <dcrowell@us.ibm.com>